product hidden on domain is moved to framework package
