+++ packages/framework/src/Model/Product/ProductDomain.php
@@ -139,6 +139,12 @@ class ProductDomain
     protected $saleExclusion;
 
     /**
+     * @var bool
+     * @ORM\Column(type="boolean")
+     */
+    protected $domainHidden;
+
+    /**
      * @param \Shopsys\FrameworkBundle\Model\Product\Product $product
      * @param int $domainId
      */
@@ -383,5 +389,21 @@ class ProductDomain
     public function setSaleExclusion(bool $saleExclusion): void
     {
         $this->saleExclusion = $saleExclusion;
+    }
+
+    /**
+     * @param bool $domainHidden
+     */
+    public function setDomainHidden(bool $domainHidden): void
+    {
+        $this->domainHidden = $domainHidden;
+    }
+
+    /**
+     * @return bool
+     */
+    public function isDomainHidden()
+    {
+        return $this->domainHidden;
     }
 }
