+++ packages/framework/src/Model/Product/ProductVisibilityRepository.php
@@ -172,6 +172,8 @@ class ProductVisibilityRepository
                                 AND pcd.domain_id = pv.domain_id
                                 AND cd.visible = TRUE
                         )
+                        AND 
+                        (pd.domain_hidden = FALSE)
                     )
                     THEN TRUE
                     ELSE FALSE
